+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+
+## master
+
+**BC break:** 
+- [\#19](https://github.com/keboola/php-component/pull/19): path when writing manifests is now relative to `out/files` (or `out/tables` respectively). 
+    - before: `$manager->writeTableManifest('/data/out/tables/table.csv', /*...*/)` 
+    - after:`$manager->writeTableManifest('table.csv', /*...*/)`
+    
+
 ## 2.1.0
 
 **Feature:**